@@ -9,23 +9,17 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.4.2",
-    "@fortawesome/free-regular-svg-icons": "^6.4.2",
-    "@fortawesome/free-solid-svg-icons": "^6.4.2",
-    "@fortawesome/vue-fontawesome": "^2.0.10",
+    "@vitejs/plugin-vue": "^4.4.0",
+    "apexcharts": "^3.46.0",
     "axios": "^1.3.5",
     "babel-runtime": "^6.26.0",
-    "bootstrap": "^4.6.0",
-    "bootstrap-vue": "^2.23.1",
-    "chart.js": "^2.9.4",
     "moment-timezone": "^0.5.43",
-    "vue": "^2.7.14",
-    "vue-chartjs": "^3.5.1",
-    "vue-search-select": "^2.9.3"
+    "vue": "^3.3.7",
+    "vue3-apexcharts": "^1.5.2",
+    "vuetify": "^3.3.23"
   },
   "devDependencies": {
     "@types/node": "^20.3.2",
-    "@vitejs/plugin-vue2": "^2.2.0",
     "sass": "^1.62.1",
     "terser": "^5.14.2",
     "ts-node": "^10.9.1",
